separate unrelated fixes and input APIs from copied snapshots
Keep this proposal focused on copied terminal state. The extracted changes are being proposed independently; the downstream release remains pinned to the existing combined fork revision.

+++ Sources/SwiftTerm/Pty.swift
@@ -155,7 +155,7 @@ public class PseudoTerminalHelpers {
         indices += [VDISCARD, VEOL2, VLNEXT, VREPRINT, VSWTC, VWERASE]
 #endif
 
-        let disabledValue = fpathconf(masterPtyDescriptor, Int32(_PC_VDISABLE))
+        let disabledValue = fpathconf(masterPtyDescriptor, _PC_VDISABLE)
         let disabledByte: UInt8? = disabledValue >= 0 && disabledValue <= 0xff
             ? UInt8(disabledValue)
             : nil
